capstone_v2/withyou_system/requirements.txt
# Dependencies
google-adk>=0.1.0
google-genai>=0.3.0
pydantic>=2.0.0
python-dotenv>=1.0.0
# For future production usage
opentelemetry-api
opentelemetry-sdk